Storybook: Showing source code examples (#126)
* Storybook: Showing source code examples

* Storybook: Code re-organization

---------
