+++ .storybook/main.ts
@@ -3,8 +3,8 @@ import type { StorybookConfig } from "@storybook/html-vite";
 const config: StorybookConfig = {
   stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
   addons: [
+    "@jls-digital/storybook-addon-code",
     "@storybook/addon-essentials",
-    "@chromatic-com/storybook",
     "@storybook/addon-interactions",
   ],
   framework: {
