+++ src/stories/experiments.stories.ts
@@ -1,41 +1,40 @@
 import type { Meta, StoryObj } from '@storybook/html'
-import { CosmosStoryProps, createCosmos } from '@/graph/stories/cosmos'
 
-import { generateMeshData } from './data.mesh'
+import { CosmosStoryProps } from './create-cosmos'
+import { MeshWithHolesStory } from './experiments/mesh-with-holes'
+import { FullMeshStory } from './experiments/full-mesh'
+
+import createCosmosRaw from './create-cosmos?raw'
+import generateMeshDataRaw from './generate-mesh-data?raw'
+import meshWithHolesRaw from './experiments/mesh-with-holes?raw'
+import fullMeshRaw from './experiments/full-mesh?raw'
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta: Meta<CosmosStoryProps> = {
   title: 'Experiments',
-  parameters: {
-    controls: {
-      disable: true,
-    },
-  },
-  render: (args) => {
-    return createCosmos(args)
-  },
+}
+
+type Story = StoryObj<CosmosStoryProps>;
+
+const sourceCodeAddonParams = [
+  { name: 'create-cosmos', code: createCosmosRaw },
+  { name: 'generate-mesh-data', code: generateMeshDataRaw },
+]
+
+export const FullMesh = FullMeshStory as Story
+FullMesh.parameters = {
+  sourceCode: [
+    { name: 'Story', code: fullMeshRaw },
+    ...sourceCodeAddonParams,
+  ],
+}
+export const MeshWithHoles = MeshWithHolesStory as Story
+MeshWithHoles.parameters = {
+  sourceCode: [
+    { name: 'Story', code: meshWithHolesRaw },
+    ...sourceCodeAddonParams,
+  ],
 }
 
 // eslint-disable-next-line import/no-default-export
 export default meta
-type Story = StoryObj<CosmosStoryProps>;
-
-// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
-const fullMeshData = generateMeshData(40, 30, 15, 1.0)
-export const FullMesh: Story = {
-  args: {
-    pointPositions: fullMeshData.pointPositions,
-    links: fullMeshData.links,
-    pointColors: fullMeshData.pointColors,
-  },
-}
-
-export const MeshWithHoles = (): HTMLDivElement => {
-  const { pointPositions, links, pointColors } = generateMeshData(40, 80, 15, 0.8)
-
-  return createCosmos({
-    pointPositions,
-    links,
-    pointColors,
-  })
-}
